UI refactoring and added conditionals for nav buttons
