--- frontend/src/app/navBar/NavBar.css
@@ -1,86 +1,0 @@
-/* Header navigation */
-.header-nav-container {
-    background-color: white;
-    z-index: 2;
-    border-bottom: 5px solid var(--neutral-dark);
-    position: -webkit-sticky;
-	position: sticky;
-	top: 0;
-}
-
-.header-nav {
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    display: flex;
-    align-items: center;
-}
-
-@media (min-width: 576px) {
-    .header-nav {
-        flex-wrap: wrap;
-        justify-content: space-between;
-    }
-}
-
-.nav-logo {
-    padding-right: 10px;
-}
-
-.nav-logo img {
-    height: 50px;
-}
-
-.nav-title {
-    font-size: 0.9rem;
-    font-weight: 600;
-    padding-right: 20px;
-    margin-right: auto;
-    display: none;
-}
-
-@media (min-width: 576px) {
-    .nav-logo img {
-        height: 70px;
-    }
-    .nav-title {
-        font-size: 1rem;
-        display: block;
-    }
-}
-
-
-/* Footer navigation */
-.footer-nav-container {
-    background-color: white;
-    z-index: 1;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-}
-
-@media (min-width: 768px) {
-    .footer-nav-container {
-        display: none;
-    }
-}
-
-.footer-nav {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    padding: 1rem 0;
-}
-
-.footer-nav i {
-    color: var(--neutral-dark);
-    font-size: 24px;
-}
-
-.footer-nav > a {
-    width: 100px;
-}
-
-.footer-nav span {
-    font-size: 0.9rem;
-}
