--- frontend/src/features/user/userNav/UserNav.css
@@ -1,60 +1,0 @@
-/* Overall */
-.user-nav {
-    position: relative;
-    background-color: white;
-    display: none;
-}
-
-@media (min-width: 768px) {
-    .user-nav {
-        display: block;
-    }
-}
-
-
-/* Toggler button */
-.user-nav-toggler {
-    box-shadow: rgba(100, 100, 111, 0.2) 0px 5px 29px 0px;
-    border-radius: 1000px;
-    height: 3rem;
-    border: 1px solid var(--neutral-light);
-    padding: 10px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.user-nav-toggler i {
-    padding: 0 5px;
-    color: var(--neutral-dark);
-}
-
-.user-nav-toggler > .fa-user {
-    font-size: 24px;
-}
-
-/* User navigation menu */
-.user-nav-content {
-    position: absolute;
-    top: 3.5rem;
-    right: 0;
-    box-shadow: rgba(100, 100, 111, 0.2) 0px 5px 29px 0px;
-    border: 1px solid var(--neutral-light);
-    border-radius: 10px;
-    background-color: inherit;
-    width: 200px;
-}
-
-.user-nav-link {
-    display: flex;
-    align-items: center;
-    background-color: inherit;
-    border: none;
-    width: 100%;
-    height: 3rem;
-    padding: 0 15px;
-}
-
-.user-nav-link:hover {
-    font-weight: 700;
-}
